2nd Boss and EventManager
Started working on the second boss, currently very rough work in progress. Realized that bullets definitely need to be cleared when the boss is defeated, so I made an EventManager to handle this, this EventManager will handle more things later on, but currently is only used to clear bullets and enemies when a boss dies.