Fixed Camera/Scene Placement, Rockets
The scene now has everything scaled properly based on screen size. Fixed some of the placement of some of the spawning stuff in the scene so that it fits better with the view. Made sure that rockets track the player. Fixed some issues with the 2nd Boss prefab not matching the 2nd Boss in the 2nd Boss scene.