Bullet Clarity Changes and Slight Audio Mixing
Changed the bullet sprite to have white in is so it's easier to see on any background, it may be a good idea to make the backgrounds generally a little darker and enemies a little brighter.

Change audio so that music is quieter and shooting sounds are slightly louder. We're missing charge laser sounds.